practice button goes to practice creen
